Move json parser utility functions to own module
